highp is entirely unnecessary and makes things slower

+++ src/client/components/visual_mesh/camera/shaders/mesh.frag
@@ -1,5 +1,5 @@
-precision highp float;
-precision highp int;
+precision lowp float;
+precision lowp int;
 
 uniform sampler2D image;
